@davidsyoung 2025-03-08
> Are you using `cmake` to build? The object file for the new file that I added (`iqk_flash_attn.cpp`) is missing from the link command. It should be automatically added with `cmake`.

Ah, I think that'll fix it. I was using the `full-cuda.Dockerfile` to run and I believe it was using a version of `make` still from previously forked `llama.cpp`.