@jukofyork 2026-02-07
> @jukofyork
> 
> Yes, I agree that this could be (should be) re-visited now with the availability of a different MLA model. But I think the estimates you are making assume performance scales linearly with batch size, which it does not, at least not when you go past 8 tokens or so. Hence, unfortunately, it is a matter of experimentation rather than theory. Me, being a theoretical physicist and still remembering how much I hated doing experiments, which were an unavoidable part of undergraduate physics education, would take theory over experiment every day of the week. When this is possible, that is.

Yeah agreed. Sadly I don't think any results from my setup will help much though - I just seem to have such weird variation from using NUMA and/or the strange alignment problem from the PCI-E bus... The smaller `GLM-4.7-Flash` model would probably help with this, but it too seems to have several oddball parameters you and mainline have run into, and not sure if those results would translate well to the more sane `Deeepseek`/`Kimi-K2` parameters :frowning: